+++ frontend/src/lib/components/settings/OnlinePersistenceCheck.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
   import { dev } from '$app/environment';
   import { environment } from '$lib/util/environment';
+  import { identifierStore } from '$lib/util/stores/store-identifier';
   import {
     resetSharedIdentifier,
     sharedIdentifierStore,
@@ -26,23 +27,15 @@
   let shareDataOnline;
   let name = '';
   let saveButton;
-  let connectButton;
   let openIdentifierInfo = false;
-  let idInput = '';
   let snackbar: Snackbar;
   let snackbarMessage = '';
 
   $: sameName = $sharedIdentifierStore.name === name;
   $: saveSubmittable = name && !sameName;
-  $: connectSubmittable = !!idInput;
   $: if (saveButton) {
     saveButton.$$set({
       disabled: !saveSubmittable,
-    });
-  }
-  $: if (connectButton) {
-    connectButton.$$set({
-      disabled: !connectSubmittable,
     });
   }
 
@@ -55,10 +48,12 @@
     }
   });
 
-  const checkIdentifierReset = async () => {
+  const triggerCloudPersistence = async () => {
     if (shareDataOnline) {
       resetSharedIdentifier();
       name = '';
+    } else {
+      connectOnlineIdentifier();
     }
   };
 
@@ -116,16 +111,21 @@
   };
 
   const connectOnlineIdentifier = async () => {
-    await fetch(`${apiURL}/identifiers/${idInput}`)
+    const loginId = $identifierStore;
+    if (!loginId) {
+      triggerSnackbar($t('page.settings.onlinePersistence.snackbarNoLoginId'));
+      return;
+    }
+
+    await fetch(`${apiURL}/identifiers/${loginId}`)
       .then((response) => response.json())
       .then((data) => {
         if (data.statusCode === 404) {
           triggerSnackbar(
             $t('page.settings.onlinePersistence.snackbarNoIdentifierFound', {
-              identifier: idInput,
+              identifier: loginId,
             }),
           );
-          idInput = '';
           return;
         }
 
@@ -141,7 +141,6 @@
             identifier: $sharedIdentifierStore.id,
           }),
         );
-        idInput = '';
       })
       .catch((error) => {
         triggerSnackbar(error);
@@ -155,25 +154,13 @@
       {$t('page.settings.onlinePersistence.title')}
     </h2>
     <p class="online_persistence_info_text">
-      {$t('page.settings.onlinePersistence.description1')}
-    </p>
-    <p class="online_persistence_info_text">
-      {$t('page.settings.onlinePersistence.description2')}
-      <b>
-        {$t('page.settings.onlinePersistence.connectionData')}
-      </b>
-
-      {$t('page.settings.onlinePersistence.description3')}
-      <b>
-        {$t('page.settings.onlinePersistence.loadConnection')}
-      </b>
-      {$t('page.settings.onlinePersistence.description4')}
+      {@html $t('page.settings.onlinePersistence.description1')}
     </p>
 
     <FormField>
       <Switch
         bind:checked={shareDataOnline}
-        on:click={checkIdentifierReset}
+        on:click={triggerCloudPersistence}
         icons={false}
         color="secondary"
       />
@@ -205,16 +192,10 @@
       </Dialog>
     {/if}
 
-    {#if shareDataOnline}
+    {#if shareDataOnline && $sharedIdentifierStore.id}
       <div class="input_area">
         <Card padded class="connection_card">
-          {#if !$sharedIdentifierStore.id}
-            <h3>
-              {$t('page.settings.onlinePersistence.establishNewConection')}
-            </h3>
-          {:else}
-            <h3>{$t('page.settings.onlinePersistence.editConnection')}</h3>
-          {/if}
+          <h3>{$t('page.settings.onlinePersistence.editConnection')}</h3>
 
           <p>
             {$t('page.settings.onlinePersistence.editConnectionDescription')}
@@ -230,11 +211,9 @@
               >
                 <Icon class="material-icons" slot="leadingIcon">badge</Icon>
                 <HelperText slot="helper"
-                  >{!$sharedIdentifierStore.id
-                    ? $t('page.settings.onlinePersistence.yourNameQuestion')
-                    : $t(
-                        'page.settings.onlinePersistence.yourNewNameQuestion',
-                      )}</HelperText
+                  >{$t(
+                    'page.settings.onlinePersistence.yourNameQuestion',
+                  )}</HelperText
                 >
               </Textfield>
             </div>
@@ -251,37 +230,6 @@
               {:else}
                 <Label>{$t('page.shared.save')}</Label>
               {/if}
-            </Button>
-          </div>
-        </Card>
-
-        <Card padded class="connection_card">
-          <h3>{$t('page.settings.onlinePersistence.loadConnection')}</h3>
-          <p>
-            {$t('page.settings.onlinePersistence.loadConnectionDescription')}
-          </p>
-
-          <div class="button_group">
-            <div>
-              <Textfield
-                bind:value={idInput}
-                label={$t('page.settings.onlinePersistence.yourId')}
-              >
-                <Icon class="material-icons" slot="leadingIcon">badge</Icon>
-                <HelperText slot="helper"
-                  >{$t(
-                    'page.settings.onlinePersistence.personalIdQuestion',
-                  )}</HelperText
-                >
-              </Textfield>
-            </div>
-
-            <Button
-              on:click={connectOnlineIdentifier}
-              bind:this={connectButton}
-            >
-              <Icon class="material-icons">cloud_download</Icon>
-              <Label>{$t('page.shared.connect')}</Label>
             </Button>
           </div>
         </Card>