+++ frontend/src/routes/settings/+page.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
   import OnlinePersistenceCheck from '$lib/components/settings/OnlinePersistenceCheck.svelte';
+  import SettingsDashboard from '$lib/components/settings/SettingsDashboard.svelte';
   import { utilityRoutes } from '$lib/config/applications';
   import { languageStore } from '$lib/util/stores/store-language';
   import { setLocale } from '$lib/util/translations';
@@ -20,6 +21,7 @@
 </svelte:head>
 
 <section>
+  <SettingsDashboard />
   <OnlinePersistenceCheck />
 </section>